demo/requirements.txt
streamlit>=1.28.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
python-dotenv>=1.0.0
typing-extensions>=4.5.0

